pyproject.toml

[build-system]
requires = ["maturin>=1.0"]
build-backend = "maturin"

[tool.maturin]
module-name = "sealwatch._sealwatch"
bindings = "pyo3"

[project]
name = "sealwatch"
version = "2025.09"
dependencies = [
  "opencv-python",
  "numpy",
  "scipy",
  "jpeglib",
  "h5py",
  "pandas",
  "pillow",
  "timm",
  "torch",
  "torchvision",
  "tqdm"
]
requires-python = ">= 3.8"
authors = [
    {name = "Martin Benes", email = "martin.benes@uibk.ac.at"},
    {name = "Benedikt Lorch", email = "benedikt.lorch@uibk.ac.at"}
]
description = "Implementation of modern image steganalysis algorithms"
readme = "README.md"
license = {file = "LICENSE"}
keywords = ["steganalysis", "image", "jpeg"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Other Audience",
    "License :: OSI Approved :: Mozilla Public License 2.0 (MPL 2.0)",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Topic :: Communications",
    "Topic :: Education",
    "Topic :: Multimedia",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Security",
    "Topic :: Utilities"
]

[project.urls]
Homepage = "https://pypi.org/project/sealwatch/"
Documentation = "https://sealwatch.readthedocs.io/"
Source = "https://github.com/uibk-uncover/sealwatch/"

[project.optional-dependencies]
dev = [
    "flake8",
    "parameterized",
    "conseal",
    "scikit-learn",
    "maturin"
]
deploy = []

[tool.coverage.report]
fail_under = 80

[tool.coverage.run]
branch = true
include = ["sealwatch/*"]
command_line = "-m unittest discover -s ."
